backend/pyproject.toml
[project]
name = "offle_assistant"
version = "0.1.0"
description = "A simple framework to build offline chatbots."
authors = [
    {name = "beatoven9",email = "reeedt@gmail.com"}
]
packages = [
    {include = "offle_assistant"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "ollama (>=0.4.5,<0.5.0)",
    "prompt-toolkit (>=3.0.48,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "qdrant-client (>=1.13.2,<2.0.0)",
    "pypandoc-binary (>=1.15,<2.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "fastapi[all] (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "redis (>=5.2.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "appdirs (>=1.4.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]

[tool.poetry.scripts]
offle_assistant = "offle_assistant.main:start"
offle_assistant_cli = "offle_assistant.cli._cli:main"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
